국제화 관련 코드 제거
