--- app/[locale]/opengraph-image.tsx
@@ -1,56 +1,0 @@
-import { ImageResponse } from "next/og";
-/*
-import { readFile } from "node:fs/promises";
-import { join } from "node:path";
-*/
-// Image metadata
-export const alt = "GalaxyChoi Portfolio";
-export const size = {
-  width: 1200,
-  height: 630,
-};
-
-export const contentType = "image/jpg";
-
-// Image generation
-export default async function Image() {
-  // Font loading, process.cwd() is Next.js project directory
-  /*
-  const interSemiBold = await readFile(
-    join(process.cwd(), "assets/Inter-SemiBold.ttf")
-  );*/
-
-  return new ImageResponse(
-    (
-      // ImageResponse JSX element
-      <div
-        style={{
-          fontSize: 128,
-          background: "black",
-          width: "100%",
-          height: "100%",
-          display: "flex",
-          alignItems: "center",
-          justifyContent: "center",
-        }}
-      >
-        GalaxyChoi Portfolio
-      </div>
-    ),
-    // ImageResponse options
-    {
-      // For convenience, we can re-use the exported opengraph-image
-      // size config to also set the ImageResponse's width and height.
-      ...size,
-      /*
-      fonts: [
-        {
-          name: "Inter",
-          data: interSemiBold,
-          style: "normal",
-          weight: 400,
-        },
-      ],*/
-    }
-  );
-}
